pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avicenna-agent"
version = "0.1.0"
description = "Avicenna: AI Agent with MCP Integration"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0"
]

[project.scripts]
avicenna = "source.avicenna.main:app"

# --- FIX BELOW: Explicitly tell setuptools what to include ---
[tool.setuptools.packages.find]
include = ["source*"]  # Only include the 'source' package
exclude = ["web*", "tests*"]   # Ignore 'web' and 'tests'